Create utils directory, refactor routes.py, move some routes.py avatar functions to utils/avatars.py
